add models

+++ db/migrate/20180514150406_create_players.rb
@@ -2,7 +2,7 @@ class CreatePlayers < ActiveRecord::Migration[5.2]
   def change
     create_table :players do |t|
       t.string :name
-      t.smallint :level
+      t.integer :level
       t.integer :health
       t.string :career
       t.references :server, foreign_key: true
