+++ db/migrate/20180514163709_create_monsters.rb
@@ -2,12 +2,12 @@ class CreateMonsters < ActiveRecord::Migration[5.2]
   def change
     create_table :monsters do |t|
       t.integer :health
-      t.smallint :level
+      t.integer :level
       t.integer :attack
 
       t.timestamps
     end
-    
+
     ActiveRecord::Base.connection.execute(<<~SQL
       alter table monsters 
       add constraint fk_monsters_creatures
