+++ db/migrate/20180514163952_create_quests.rb
@@ -3,7 +3,7 @@ class CreateQuests < ActiveRecord::Migration[5.2]
     create_table :quests do |t|
       t.string :name
       t.boolean :is_main
-      t.smallint :level
+      t.integer :level
       t.references :npc, foreign_key: true
 
       t.timestamps
